Fix warning.
git-svn-id: https://src.develer.com/svnoss/bertos/trunk@997 38d2e660-2303-0410-9eaa-f027e97ec537

+++ kern/monitor.c
@@ -158,7 +158,7 @@ static void NORETURN monitor(void)
 
 			if (free < 0x20)
 				kprintf("MONITOR: WARNING: Free stack for process '%s' is only %x chars\n",
-						p->monitor.name, free);
+						p->monitor.name, (unsigned int)free);
 
 			timer_delay(500);
 		}
